Split serialize_parse implementations into multiple headers
