+++ src/middleware/log/dccl_log_plugin.h
@@ -25,6 +25,8 @@
 
 #include <dccl/codec.h>
 
+#include "goby/middleware/serialize_parse_dccl.h"
+
 #include "protobuf_log_plugin.h"
 
 namespace goby
