+++ src/test/zeromq/single_thread_app1/test.cpp
@@ -19,6 +19,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Goby.  If not, see <http://www.gnu.org/licenses/>.
 
+#include "goby/middleware/serialize_parse_protobuf.h"
 #include "goby/time.h"
 #include "goby/time/io.h"
 #include "goby/zeromq/single-thread-application.h"
