+++ src/test/zeromq/zeromq_portal_without_interthread/test.cpp
@@ -25,6 +25,7 @@
 #include <atomic>
 #include <deque>
 
+#include "goby/middleware/serialize_parse_protobuf.h"
 #include "goby/middleware/transport.h"
 #include "goby/zeromq/transport-interprocess.h"
